Implement Quantitative Financial ufuncs fv, pv, npv, irr
 - Implement fv, pv, npv, and irr using FFI-accelerated binary/ternary operations.
- Statically type financial functions to use NDArray<Float64>.
- Fix 0D ternary broadcasting in C kernels.
- Fix generic type erasure in _wrapScalar.
- Rename comparison operators to camelCase (equal, greaterEqual, etc.) and use _compareHelper with FFI ndarray_compare binding.
- Optimize _getStrippedCoeffs and _hasSameSign to use zero-copy slice and FFI nonzero instead of copying to Dart lists.
- Add comments explaining performance trade-offs for manual loop in _hasSameSign.
- Update tests and examples to use NDArray<Float64> and NDArray.scope instead of manual dispose.

TAG=agy
CONV=833eb8e3-525e-4f24-86f3-37fada327320

+++ pkgs/ndarray/test/hypot_power_test.dart
@@ -46,9 +46,9 @@ void main() {
       final a = NDArray.fromList(Int32List.fromList([2, 3]), [2], DType.int32);
       final b = NDArray.fromList(Int32List.fromList([2]), [1], DType.int32);
       final p = power(a, b);
-      expect(p.data[0], 4.0);
-      expect(p.data[1], 9.0);
-      expect(p.dtype, DType.float64);
+      expect(p.data[0], 4);
+      expect(p.data[1], 9);
+      expect(p.dtype, DType.int32);
     });
   });
 }
